refactor: cqrs 패턴 사용하지 않고 하나의 controller와 service에서 모두 데이터 조회, 변경 처리
